+++ backend/app/models/__init__.py
@@ -7,8 +7,8 @@ helper of 'from app import models' or import app.models
 loads/exports models
 
 '''
-
-from app.models.user import User
+from app.models.admin_user import AdminUser
+#from app.models.user import User # [info] user foreignkeyerror
 from app.models.customers import Customer
 from app.models.products import Product, ProductVariant
 from app.models.invoice import Invoice, InvoiceItem, InvoiceTax, InvoiceStatus