+++ backend/app/models/orders.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 
-# filename: orders.py
+# filename: /backend/app/models/orders.py
 # descr: relationship with product, user
 from __future__ import annotations
 from typing import TYPE_CHECKING
 from decimal import Decimal
 from sqlite3.dbapi2 import Timestamp
-
+from sqlalchemy.orm import relationship
 from app.extensions import (
     db,
     Mapped,
@@ -29,7 +29,7 @@ if TYPE_CHECKING:
     from .customers import Customer
 
 class Order(TimeStampModel):
-    __tablename__ = 'orders'
+    __tablename__ = "orders"
     "table arguments is a class attribute"
     "This attribute accommodates both positional as well as keyword arguments that are normally sent to the Table constructor. "
     __table_args__ = (
@@ -66,31 +66,15 @@ class Order(TimeStampModel):
         db.ForeignKey('products.id')
     )
     
-    product: Mapped["Product"] = db.relationship(
-        "Product",
-        back_populates="orders"
-    )
-    
     created_at: Mapped[datetime] = mapped_column(
         DateTime,
         nullable=False,
         default=lambda: datetime.now(timezone.utc),
     )
 
-    created_by_user: Mapped["User"] = db.relationship(
-        "User",
-        back_populates="orders"
-    )
-
-    customer: Mapped["Customer"] = db.relationship(
-        "Customer",
-        back_populates="orders"
-    )
-
-
     created_by_user_id: Mapped[int] = mapped_column(
         BigInteger,
-        db.ForeignKey('users.id'),
+        db.ForeignKey('adminUsers.id'),
         nullable=False
     )
     
@@ -98,4 +82,17 @@ class Order(TimeStampModel):
         BigInteger,
         db.ForeignKey("customers.id"),
         nullable=False
+    )
+
+    product: Mapped["Product"] = db.relationship(
+        "Product",
+        back_populates="orders"
+    )
+    created_by_user: Mapped["AdminUser"] = db.relationship(
+        "AdminUser",
+        back_populates="orders"
+    )
+    customer: Mapped["Customer"] = db.relationship(
+        "Customer",
+        back_populates="orders"
     )