+++ CHANGELOG.md
@@ -6,10 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [1.1.0](https://github.com/gutiluis/ERP_cloud/compare/v1.0.0...v1.1.0) (2026-06-19)
 
 
-### ### Added
+### Added
 
-* add /api/admin/products endpoint ([6a42af1](https://github.com/gutiluis/ERP_cloud/commit/6a42af1d51f41bd80fbe28fff238fe01d7112ab2))
-* **admin/customers:** implemente flasgger documentation for customer details ([4c239db](https://github.com/gutiluis/ERP_cloud/commit/4c239db92b60d19600cd66b2a425f040863641b4))
+* **admin/customers:** implemented flasgger documentation for customer details 
 
 ## 1.0.0 (2026-06-19)